+++ packages/v1/sdk-js/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @copilotkit/sdk-js
 
+## 1.53.0
+
+### Patch Changes
+
+- Updated dependencies [1510f64]
+- Updated dependencies [bf1fc6f]
+  - @copilotkit/shared@1.53.0
+
+## 1.53.0-next.6
+
+### Patch Changes
+
+- @copilotkit/shared@1.53.0-next.6
+
+## 1.53.0-next.5
+
+### Patch Changes
+
+- @copilotkit/shared@1.53.0-next.5
+
+## 1.52.2-next.4
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.2-next.4
+
+## 1.52.2-next.3
+
+### Patch Changes
+
+- Updated dependencies [1510f64]
+  - @copilotkit/shared@1.52.2-next.3
+
 ## 1.52.2-next.2
 
 ### Patch Changes