+++ packages/v2/eslint-config/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @copilotkitnext/eslint-config
 
+## 1.53.0
+
+## 1.53.0-next.6
+
+## 1.53.0-next.5
+
+## 1.52.2-next.4
+
+## 1.52.2-next.3
+
 ## 1.52.2-next.2
 
 ## 1.52.2-next.1